pyproject.toml
[project]
name = "assessment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.13.0",
    "mcp>=1.27.0",
    "openai>=2.33.0",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
]
